[config] move execution_config to own file
In addition, I also made the genesis_file_location a PathBuf instead of
a string
